Add a function to print to the center of a screen
"void printCenter(WINDOW *winName, char message[])" takes a given
message and outputs it to the center of a given ncurses window.

The function is placed in the new namespace IO which will handle all
input and output for the program. Any interaction with the user must go
through IO.

All files for the new namespace (IO.cpp and IO.hpp) have been added to
the repository in their proper directories.