Lengthen README.md
Edited README.md to make it a full readme file written in Markdown. It
is still a work in progress (like the rest of this project).

A few major updates are on their way, I just need to work on some
functions for integer overflow and finish implementing a character
class.