Remove redundancy and errors in overflow functions
The overflowAdd and overflowSubtract functions for unsigned integers
had an if statement in them that checked if they were given negative
parameters. They can only take unsigned (positive) integers. This made
the if statment useless and it is now removed.

The overflowSubtract function for unsigned integers takes a signed
modifier instead of an unsigned one. This was fixed.

The implementation for overflowDivide will come soon (by the end of the
week).