Add divison implementation
I finally finished the general namespace. Implementing the divison
felt too easy, so there might be an error or two. If I find one I'll
fix it.

A character class will probally be the next major update, either that
or programs to test all of the functions (the ones in general and io).

There are a bunch of minor problems with the repository that will also
be fixed soon, including making the line endings consistent
(LF vs CRLF) and adding a CONTRIBUTORS file so I can accept
contributions to the repository.