Fix implementation for overflowDivide
When I first implemented I forgot that dividing two negative numbers
can make a bigger number (-25 / -5 = 5), so I added an if statement to
check if the number goes over maxNum. If it does, then the function
will return maxNum.

The next update will either fix the line endings (convert to LF in the
repository) or add a CONTRIBUTORS.txt file to tell potential
contributors how to contribute to the repository.