Fix minor bugs in character class implementation
VERY minor commit today, just two tiny bugfixes in the new character
class. A function call was spelled wrong
(overflowSubtract vs overflowSUbtract) and I fixed the changeLevel()
function so that it didn't take parameters. It does not need to take
parameters, so why bother?

I have written most of the test code for overflowAdd in the general
namespace. This will be made into a commit tomorrow.